requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
Pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
